Fix dynamic_inventory.py test bug
dynamic_inventory.py implements a test to check for
duplicate hosts assigned to the same IP address
in user supplied inventory.

A test has been written which asserts the output of
the error raised by MultipleHostsWithOneIPError to
equal a string in which two hostnames appear.  Since
the inventory created in dynamic_inventory is a python
dictionary, the ordering of the elements cannot be
guarenteed.  This will cause an occasional failure
during automated testing.

This patch creates a small list of the hostnames
supplied to MultipleHostsWithOneIPError, sorts the
list, and supplies the list index values in place
of the variable names directly.  This will ensure
that our exception text will always match our
assertion during testing.

Change-Id: I43b3c637a3944efb705b5c4f495ab3dba1faf893