Automatically increment the patch ver with sources-branch-updater
With each SHA bump the release patch version needs to be incremented
to ensure that the repo build process uses the correct folder name.

This implements an automated update of the appropriate file to make
that happen when the sources-branch-updater is executed.

Change-Id: Ide700f1c474adb9dd4bbe7a5f8d541e5e21d9fcc
(cherry picked from commit 87e32dcfc87eacf3f1eae1059d8651f5ae3e56ba)