Isolate Ansible bootstrap from repo servers
On first deployment bootstrap-ansible.sh pulls its packages from PyPI.
On upgrade, when the deployment host is shared with a controller, a
pip.conf file exists to restrict pip to the packages hosted by the repo
containers. This causes bootstrap-ansible.sh to fail when the repo
server created by the previous version does not contain a compatible
version.

This commit adds the flag '--isolated' to allow the bootstrap process to
bypass the repo server.

To allow the pip command to fall back on using the isolated flag the
code is refactored to determine the version of pip prior to its use.

Change-Id: Ie9834eb55a689d6f74441cc78e34b62be5d82141
(cherry picked from commit 6be15b8dddd765edcb06cfe87d2cc72bc8567d5f)