install rabbitmq-server in serial
In order to prevent race conditions with nodes joining the cluster
simultaneously when the cluster is first formed, we move the rabbitmq
installation play to be 'serial: 1'. However, when the nodes are being
upgraded, it cannot be done in serial so in this case we set 'serial: 0'

There are some tasks/roles called in this playbook that can still be
run in parallel, so we split out the rabbitmq-server install into a
separate play so that we only serialise the parts that are necessary
to ensure maximum efficiency.

Based on commit: 21c24789019bf1ee85347356fe6c70b0b9045d7c

Change-Id: I97cdae27fdce4f400492c2134b6589b55fbc5a61
Fixes-Bug: #1573030