DOCS: Configuration section - cleanup
As per discussion in the OSA docs summit session, clean up
of installation guide. This fixes typos, minor RST mark up
changes, and passive voice.

This patch also merges a some of the sections into the larger
chapter. This is in an effort to remove multiple smaller
files.

This patch is the first of many to avoid major conflicts.

Combined backport of:
 - I38daa515ba47fde7719cd0bd3e0e40c2cd0f39f1
   (cherry picked from commit 71554ca7cf7447ce2c6e114ff1c11e2c2d42a6f0)
 - I8b5a16fef83283b5dda535a279c04dad52a7fcc8
   (cherry picked from commit 59694c78828f42a4144d746523ad2e0a17e5c2e5)

Change-Id: I38daa515ba47fde7719cd0bd3e0e40c2cd0f39f1