Include reseller admin role correctly
The name of the reseller_admin_role in the proxy config needs
to map to the correct role within the deploy. It was specified
manually and inconsistently though a variable exists in the role
defaults.

Backported by community request.
A cherry pick is not possible due to IRR in Mitaka.

Note that the Keystone catalog has to be adjusted to remove the V3
endpoint for Swift, replacing it with a V2 endpoint because
Ceilometer does not support V3 auth in Liberty (fixed in Mitaka)
before this fix can be used.

Change-Id: Iacf45828cff5907bf204c35e1d234faacffeab65
Backport-Of: Ib93fc57f1e26e8a0943a7bd7c2212ce94be54f7c
Related-Bug: #1542391