Create ceph python library symlinks
Currently, os_nova, os_cinder, and os_glance roles all install
python-cephlibs, which is a python package Hugh Saunders created for
Ceph Hammer python libraries.  The virtualenvs created in these roles
need access to these python libraries and by using Hugh's python
package we could simply pip install into the virtualenv.

This has been working fine, but makes the assumption that the deployer
is using Ceph Hammer, which may not be true.  Also, later versions of
Ceph have python C extensions, which may be more difficult to package.

This commit simply leverages the existing Ceph libraries installed
outside of the venv by checking for the path of the library using
python itself.

NOTE: I would prefer to use a None instead of '' in the ternaries,
but seem to hit issues doing that.

Combined backport of:
- https://review.openstack.org/#/c/317901/
- https://review.openstack.org/#/c/339619/

Change-Id: Ib26c84f6671523b40e61f043713db58d8a756dc9