Docs: Add note about slow galera recovery
Rebuilding a galera container is stressful enough and deployers
may find that restarting the MariaDB daemon on a re-created
container takes a lot of time.

This patch adds a note to advise deployers about the time required to
sync data from the older nodes to the new, empty node. This will hopefully
restore some sanity and alleviate some stress.

Change-Id: Ife5ffed913e7ca21f7d003b7f09b740e1f678a21