Update openstack-ansible-security role SHA
The role openstack-ansible-security has been updated to fix several
bugs. This commit updates the version used by openstack-ansible to the
head of the liberty branch.

This commit switches from using a tag to using a SHA to allow future
updates of the version consumed to be done as part of the standard
release process by which all other SHAs are updated.

Change-Id: I812aa3642a7f62761a70006a4e5d312f4b7bad2c