Add tests for the _net_address_search function
The _net_address_search function looks through the provider networks and
sets a key to True if it's the main network. This commit adds tests
around the conditionals found in the function. In practice, the keys
that are used (is_ssh_address and is_container_address) are specified in
the gate user_config files, but this function seems to be there to catch
and correct errors.

The 'else' clause from the 'for' loop was moved up, since logically it
was the action that happened when the target key was not present. The
'else' clause made it re-loop, which was unnecessary.

Change-Id: Ic5920f19c8e21e324aee9c5e11831fd2c2e2b21d