Set AIO to use an OpenStack-Infra wheel mirror
To speed up an AIO build when *not* executed in the OpenStack-CI gate,
this patch determines the fastest available wheel mirror and includes
it as an extra pip index.

This speeds up the repo build process by ensuring that the wheels do
not need to be built locally, but can simply be downloaded from the
mirror.

The patch also includes some clean up which was left out of
I73fa1db5210f02d6df9dc324f8b4ec21232d06ba relating to apt conf files.

Change-Id: I56f7c8520cc6cec74df5ca8611d988039184efbb