Reduce reliance on global state for testing
In writing other tests, the INVENTORY_SKEL dictionary was mutated
globally and affecting tests. This change makes sure that it is copied
on each run, instead of re-used. During normal usage, this behavior is
not dramatically different, because the INVENTORY_SKEL dictionary is
destroyed when the script exits. However, when running tests, the module
and thus the variable stay loaded and are not implicitly destroyed.

This changed revealed that USED_IPS was also not being properly cleaned
on the duplicate IP test. In order to minimize the scope of this commit, the
affected test is fixed rather than all instances of USED_IPS access.

Change-Id: I6d1ea3447079e2f61d2d4a99ffdbb7fe311e4406