DOC New Appendix - custom component layouts
Adding a new Appendix to the Install Guide.
There are a few frequently asked questions this seeks to address:
- how to deploy on_metal. This is mentioned and demonstrated but we
still get questions, so this lays it out more thoroughly.
- how to omit a service from a deploy.
- how to change where stuff is installed / how to add a new service
to OpenStack-Ansible.

This appendix should be useful for developers as well as for
deployers so the example of using dedicated DB hardware is given
lots of words to explain the how and the why to hopefully explain
better the model for mapping inventory groups from a playbook to a
set of physical hosts.

Change-Id: I1e8dd71e009457eba574a93211e106890ebae2ef