Create symlink for libvirt save directory
Nova can fill up the partition that /var/lib/libvirt/qemu/save is on
when nova calls `image-create` on an instance. This occurs because nova
calls libvirt with managedsave which creates a temporary file in the
/var/lib/libvirt/qemu/save location. Libvirt does not allow the location
of /var/lib/libvirt/qemu/save to be configurable so we have to manage it
ourselves via bindmounts or symlinks. In this commit we remove the
/var/lib/libvirt/qemu/save directory and create a symlink in it's place,
pointing to `{{ nova_system_home_folder }}/save` instead (which we also
create). This location was chosen as deployers often have more space
here for nova instances.

Change-Id: I5c96f70aa355cc02379fe7e0681416eca53447eb
Partial-Bug: #1585325