Do not use cpu_map_update.py anymore
Because of the change [1], the file tools/cpu_map_update.py was removed,
and we rely on this file for gating.

However, according to the same review, this shouldn't be done anymore
because "nova has been fixed" (I've got no evidence of it yet).

This should fix the gate issue.

[1]: https://review.openstack.org/#/c/324562/

Change-Id: I83e3fc349a0448e0411f369bdf4888ff81cb37a8
Depends-On: I198db581db22655dc8791ba40a77a39b362c1b73
Closes-Bug: 1589440
Signed-off-by: Jean-Philippe Evrard <jean-philippe@evrard.me>