Remove the AIO metadata checksum fix from run-playbooks
With the implementation of I008bfdb2960800845703e721b38640b7434d1404
it should be safe to remove the AIO metadata checksum fix from
run-playbooks.

This patch implements the appropriate AIO configuration to ensure that
the fix is still implemented by the os_neutron role, and also implements
a convenience fix for repeated runs of run-playbooks (mostly as a proof
of concept right now).

The intent is still to wholly replace run-playbooks with the standard
production execution of the three high-level playbooks. This will be
implemented in a follow-on patch.

Change-Id: If75911c2aafa691e147f3b444d922f1dfcf87860