Remove the AIO metadata checksum fix from run-playbooks
With the implementation of I008bfdb2960800845703e721b38640b7434d1404
it should be safe to remove the AIO metadata checksum fix from
run-playbooks.

This patch implements the appropriate AIO configuration to ensure that
the fix is still implemented by the os_neutron role, and also implements
a convenience fix for repeated runs of run-playbooks (mostly as a proof
of concept right now).

The intent is still to wholly replace run-playbooks with the standard
production execution of the three high-level playbooks. This will be
implemented in a follow-on patch.

Backport note:
This is being backported due to the numerous IRC conversations which
we have related to run-playbooks being used multiple times.
Implementing this backport will reduce the number of times we have to
field questions about lost container connectivity.

Depends-On: I008bfdb2960800845703e721b38640b7434d1404
Change-Id: If75911c2aafa691e147f3b444d922f1dfcf87860
(cherry picked from commit 1133ea8de25517321068ec77980f264fdce1d073)