Temporarily disable UCA usage in OpenStack-CI
OpenStack-CI does not have Ubuntu Cloud Archive (UCA) mirrors at this stage.
The implementation of Ic1e7619b48a9b12ba876f83507497cf1e557223a set the
usage of UCA for nova to be the default, which is good for production
environments.

This patch temporarily disables its usage in OpenStack-CI while the mirror
of UCA is arranged in OpenStack-Infra.

Change-Id: I0aec0447e1faf5ae21ce542945082079d3a59727