Actually remove Ironic container creation from AIO preparation
Currently the Ironic containers are created as part of the AIO
preparation, but Ironic is not deployed due to run-playbooks.sh
excluding it by the default of 'DEPLOY_IRONIC=no'.

In order to prepare for the change of run-playbooks from a bash
script to instead using the standard deployment playbooks which
are used in the install guide, this patch removes the AIO conf.d
entry implementation for Ironic.

The example file remains in place for any deployers wishing to
test Ironic.

Change Idec60e8b85070aaa122a361141aa25b4acc152c2 intended to
achieve this goal, but mistakenly removed the ironic env.d file
from the bootstrap process instead. This patch also reverts that
change.

Change-Id: I72c13e5c81e64d66ad819fd153469224be011838