Use correct keystone auth parameters
Neutron Metadata Agent uses the authentication parameters as a fallback
method in the event that communication with Neutron API over RPC fails.
In the fallback case, it looks for Keystone v2.0 authentication
credentials but looks for them under the names:

- admin_user
- admin_password
- admin_tenant_name

Which can be the service user information. We were previously
configuring Keystone v3 authentication parameters in this space and this
causes 400 Bad Request responses from Keystone (since the Agent has null
values for those config options).

Further, the metadata agent does not need to use the admin URL to
authenticate. Instead, it can use the internal URL to retrieve its token
and then authenticate to Neutron API over HTTP.

Change-Id: Ib413d3f3f3351bef29b0e68a2cfb96b7f3dff3c3
Closes-bug: 1591282
Closes-bug: 1590957