Implement overlayfs as the backing store for the AIO
Since overlayfs is now part of the mainstream kernel, it makes sense
to take advantage of it to speed up the container creation process.

While on Ubuntu Trusty (kernel < 3.14) this is not a great idea for
production, the pre-mainline implementation of overlayfs should be
fine for the AIO use-case as it greatly improves the setup time for
the AIO and none of the container data in an AIO is anything we
need to care about too deeply (the data we care about is bind-mounted
to the host anyway).

Depends-On: I20123b83af10c0890f4d5945b9fb230acd664213
Depends-On: If5b11ef6d94fe863a75efd174957ad43b9d2e030
Change-Id: I38d387c8fe51d2d961e928ec3a6924252d17d81d