Add a rabbitmq restart in the upgrade process
During the upgrade process, a container hostname change is done.
A workaround already applied changes the hosts file to make system work
with the multiple names, but rabbitmq still complains to be unable to
connect to its peers. This should solve it: After the change in the hosts,
a restart will apply the change of hostname to the container, and it will
restart and connect to its peers.

Change-Id: I85a545274f0370bd4133ca9c823ed729ae9d717f
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>