Gate: Restrict Ansible fact gathering to base subset
In Ansible 2.x the fact gathering process includes facts gathered
by facter [1] and ohai [2]if they are available. While this may be
useful in some environment, OpenStack-Ansible does not make use of
any of these facts in any of its playbooks/roles and the facts
gathered are therefore just an extra overhead.

This patch removes the facter and aohai gathering subsets. On a
test AIO this reduced the execution time for fact gathering to the
host and all running containers from 1m5s to 6s.

[1] https://docs.puppet.com/facter/
[2] https://docs.chef.io/ohai.html

Change-Id: Id2a932a98ebfa2534bcce7ba488906fa4f03d0a6