Fix rabbitmq restart on non-rabbitmq accessible nodes
During the upgrade playbook, rabbitmq is restarted. The deploy
node will try to contact rabbitmq directly, based on its
inventory hostname. If the inventory hostname is unknown to the
deployment node, the playbook will fail.

The rabbitmq playbook now delegates the rabbitmq check to the
physical host having the rabbitmq node, on the IP used for
SSH'ing, instead of the hostname.

Closes-Bug: 1586148
Change-Id: I6b267a1fca8e894142d2a0a5de69d5ee3d333875