Remove _net_address_search from dynamic inventory
This function was mutating user-specified configuration
for the provider_networks list. Rather than change the config
the user provides, we should instead error when what is provided
does not meet our preconditions.

Now, if the provider network configuration for the management
network is missing the keys 'is_container_address' or 'is_ssh_address'
or the value of those keys is false, the dynamic inventory will exit
with a failure and a helpful message indicating the configuration issue.

Test cases associated with the function have also been removed.

Change-Id: I7d3a6229bef35e8e057a3b95ed5bfc422c1a14a3