Enable OpenStack-Infra Ubuntu Cloud Archive mirror
With the implementation of https://review.openstack.org/330152
OpenStack-Infra now has its own mirror of Ubuntu Cloud Archive.

This patch determines whether the bootstrap is running on
OpenStack-CI, and if it is it can figure out the appropriate
URL to use to access UCA and it configures the right override
to make sure that happens.

Change-Id: Ie8acab60c885454da65eda085a9c012e0af11807