Removed the default pip install options from upgrade.sh
Instead of forcing to reinstall pip packages during an upgrade, we
should instead make sure the environment always have the latest pip
package version.

This commit is part of a group of commits to ensure latest pip packages
are installed in the venv (cf. topic pip-install-latest on gerrit) for
the standard playbooks, and remove the options used by default in the
upgrade script (pip_install_options=--force-reinstall)

Closes-Bug: 1596620

Change-Id: I9ce478217d806fe04a2fc25be4cd42a875a6ffa2
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>