Adding requests to bootstrap ansible
The requests package is used on the base system and needs to be installed
early. This package is eventually installed when the lxc-hosts play is
executed which is pulled in via the "lxc-template" dependency. However due
to recent changes in the "human_log" plugin as found here [0] we need to
have this package installed sooner.

[0] - https://github.com/openstack/openstack-ansible-plugins/commit/aa69b10a1a625249696389742353432cb93f13c6

Change-Id: Id9475134950f329f777144dedb0e7c2c910ef009
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>