Added git package to the utility container
This package is needed on utility containers to allow git clones
from the repo server and/or other locations. Without this package
Tempest role can not execute because the role assumes git is already
available within an environment where it is to be executed.

Its worth noting that this works fine in Ubuntu 14.04 because of the
heavy nature of the base container image while in Ubuntu 16.04 the
container images are far more minimal.

Change-Id: I63f051cc7fa6045eed8332713de3eea241f17504
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>