Remove excessive tags
This change simply removes all of the named tags being used within the
repo-build playbook. Because the lookup needs to be executed everytime
the playbook is called it makes a lot more sense for this playbook to
use the "always" tag instead of a mix of all of the above.

Change-Id: I1ba96a731f22e65045370f43656b743943fed365
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>