Install Ansible from pypi instead of from a git clone
This patch removes the git clone machinary and sets the Ansible
installation to happen from pypi by default. The ANSIBLE_PACKAGE
environment variable can still be used to change the install to
use a git source if the deployer wishes to do so.

This change is implemented to remove another moving part from
the Ansible bootstrap process, allowing a pypi/wheel mirror to
be used instead.

Change-Id: I137956530a11b0b23adb41f6b38dd55cde892221