Support other architectures in apt sources.list
The install-apt task of the bootstrap-host role obtains the apt
repository URL by performing an awk regex to match on "/ubuntu/".

Other architectures however, such as ppc64le, contain "/ubuntu-ports/"
in the URL which causes the regex to fail.

This patch modifies the awk regex to match any characters which exist
after "/ubuntu" and before the trailing "/".

Change-Id: I5417498dc89407edd642fd8f8fdce9745acb1759