Move LXC AppArmor profile setting to the inventory
To reduce the number of times a container is restarted during the
build process, this patch moves the AA profile definition from
the service playbooks to the inventory.

This ensures that the profile can be applied when the container
is created, rather than after the fact. This, in turn, eliminates
the need to restart the container again and wait for SSH
connectivity to be restored.

There are additional configurations in some playbooks which will
require an additional facility in the lxc_container_create role
to allow something similar for container command execution. The
changes to implement this will follow in another patch once that
facility is available.

Change-Id: I688640cf06a7ed2554210e49275cc3de10371a23
Depends-On: I4c95a231af3306c5010b168e7cc181deecf1a58d