Disable ansible retry files
Currently the ansible retry files generated on failure are being saved
in the playbooks execution directory which is the new default in Ansible
"2.x". Because retry files generally do not work due to tag and task
execution issues that can come from over limiting and the fact that they
pollute the git tree the ansible retry files option has been disabled.

Change-Id: Idb4e0bd57aa2f7d6dff0277f2ffcc53905638295
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>