Dynamically determine whether ceilometer should be enabled
Currently each service requires a var to be set to enable the
configuration of ceilometer for the service.

This patch makes this setting dynamic based on whether there
are any hosts in the ceilometer_all host group.

This reduces the amount of configuration required for an
environment, making OSA simpler to use.

The 'rabbitmq_servers' group_var was previously a list of
'host:port' combinations. It is now only a set of hosts.

The new 'transport_url' configuration option for oslo_messaging
is not working with SSL enabled, so SSL transport for RabbitMQ
has been temporarily disabled.

Change-Id: Ia5ba4e26b9adcfd923a93c6ea5077cbec0c4478d