Restore telemetry service deployment
In https://review.openstack.org/344345 the telemetry deployment was
removed in order to reduce the deployment time, thus improving the
chance of success for the commit.

This patch restores the telemetry service deployment. In order to
do that, new values are needed to address undefined new variables in
the ceilometer role.

Change-Id: I4d6323b3f57a67f1189d79e01aedee267f496b7c