Fix config entry regression introduced in container restart patch
When writing the lxc container configuration files the line
needs to be normalized such that it is "key = value" If the line
is not normalized it could cause a misconfiguration of a
container, an inadvertant restart, or break the container entirely
until the config issues is resolved. This change ensures that all
configuration changes made to a container are normalized.

Fixes regression introduced in https://review.openstack.org/344834

Change-Id: I0fe3886ba810992fc08998fe5baaa27649fe1d1f
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>