[DOCS] Fix up validation failures
Attempting to build the documentation with "tox -e docs" results in a
failure due to several lint errors. This patch fixes those errors so
that the docs build successfully.

The 'docs' target is also added to the 'linters' target chain so that
the doc8 linting is executed as part of the lint checks to ensure that
these issues don't re-occur.

Change-Id: I9e1bf2cabe9a0591963d1afadbeb24cb614e0b27