Add discovery and build for multiple CPU architectures
This adds support for discovery of repo build servers for multiple
CPU architectures. When multiple architectures are detected among the
repo servers, this patch allows repo-install to assign and store a list
of 'master' build servers for each CPU architecture, then use this to
build artifacts independently for each CPU architecture in the deployment.

Partially-Implements: bp/multi-arch-repo

Change-Id: Ibfb77cc4ce87c9064f9af84946c5a9bbba5e1fb4