Confirm container data destroys
Currently running the lxc-containers-destroy playbook will, when
agreed to, destroy all containers and container data across the
entire deployment.

This patch separates the removal of the container from the removal
of the data and includes a prompt to the user to confirm each
option.

The option can be pre-answered through the use of the boolean
variable 'force_containers_data_destroy'.

Change-Id: I70c4819dbe6fc996dd780f205f3e08c300f18591