If /var/log/lxc exists, move it to the log aggregation parent
In https://review.openstack.org/344721 the log aggregation link
was added for lxc logs to ensure that the logs are collected
in a central place.

This patch ensures that any pre-existing folder (which may exist
due to the playbook being run when upgrading) is moved to the
right place. This ensures that previous logs are kept and that
upgrades execute without failure.

Change-Id: I35a9b3176a79caf9cee177442fa6b3d79e057ea9